+++ src/main/java/io/github/MeiNic/MenschAergereDichNicht/dice/LaPlaceDice.java
@@ -14,19 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package io.github.MeiNic.MenschAergereDichNicht;
+package io.github.MeiNic.MenschAergereDichNicht.dice;
 
-interface Player {
-    String getName();
-    int getPlayerState();
-    int getPlayerIndex();
-    default int getIndexOfFirstFigure() {
-        return 4 * getPlayerIndex();
-    }
-    default int getIndexOfLastFigure() {
-        return 4 + getIndexOfFirstFigure();
-    }
-    default int getIndexOfStartField () {
-        return 10 * getPlayerIndex();
+import java.util.Random;
+
+public class LaPlaceDice implements Dice {
+    private static final Random rand = new Random();
+
+    public int roll() {
+        return 1 + rand.nextInt(6);
     }
 }
