+++ src/main/java/io/github/MeiNic/MenschAergereDichNicht/figure/FigureState.java
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package io.github.MeiNic.MenschAergereDichNicht;
+package io.github.MeiNic.MenschAergereDichNicht.figure;
 
 public enum FigureState {
     IN_BASE,
