+++ src/main/java/io/github/MeiNic/MenschAergereDichNicht/gui/ImageTextPanel.java
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package io.github.MeiNic.MenschAergereDichNicht;
+package io.github.MeiNic.MenschAergereDichNicht.gui;
 
 import javax.swing.*;
 import java.awt.*;
@@ -43,8 +43,8 @@ public class ImageTextPanel extends JPanel {
         repaint();
     }
 
-    public void setImage(String name){
-	defaultImg = Resources.loadBufferedImage(name);
+    public void setImage(String name) {
+        defaultImg = Resources.loadBufferedImage(name);
         repaint();
     }
