+++ src/main/java/io/github/MeiNic/MenschAergereDichNicht/player/Dummy.java
@@ -14,15 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package io.github.MeiNic.MenschAergereDichNicht;
+package io.github.MeiNic.MenschAergereDichNicht.player;
 
-class Bot implements Player {
+public class Dummy implements Player {
     private final String name;
     private final int playerIndex;
 
-    public Bot(String name, int playerIndex) {
-        this.name = name;
-        this.playerIndex = playerIndex;
+    public Dummy(String name, int playerIndex) {
+	this.name = name;
+	this.playerIndex = playerIndex;
     }
 
     public String getName() {
@@ -30,7 +30,7 @@ class Bot implements Player {
     }
 
     public int getPlayerState() {
-        return 1;
+        return -1;
     }
 
     public int getPlayerIndex() {
